+++ ui/components/FloatingActionButton.jsx
@@ -63,6 +63,60 @@ FloatingActionButton = class FloatingActionButton extends React.Component {
             	content:<FacilityViewEdit item={newItem} />
             });
         })
+    }
+
+    createNewComplianceRule(newRule) {
+        console.log(newRule);
+        var facility = newRule.facility;
+        if(facility) {
+            var services = facility.servicesRequired;
+            //get index of the selected service
+            var idx=-1;
+            for(var i in services) {
+                if(services[i].name==newRule.service.name) {
+                    idx = i;
+                    break;
+                }
+            }
+            if(idx>=0) {
+                var service = services[idx];
+                console.log({service,idx});
+                if(!service.data) {
+                    service.data = {};
+                }
+                if(!service.data.complianceRules) {
+                    service.data.complianceRules = [];
+                }
+                //to avoid circular search remove facility and then add with just name and _id
+                var copy = _.omit(newRule,'facility','service','event');
+                if(newRule.facility) {
+	                copy.facility = _.pick(newRule.facility,'name','_id');
+	            }
+                if(newRule.event) {
+	                copy.event = _.pick(newRule.event,'name','_id');
+	            }
+	            if(newRule.service) {
+	                copy.service = _.pick(newRule.service,'name');
+	            }
+                //console.log(copy);
+                service.data.complianceRules.push(copy);
+                services[idx] = service;
+            }
+            facility.setServicesRequired(services);
+        }
+        Modal.hide();
+    }
+
+    handleCreateComplianceRuleClick() {
+        Modal.show({
+            content:<AutoForm 
+                item={{
+                    facility:Session.getSelectedFacility()
+                }}
+                schema={ComplianceRuleSchema} 
+                onSubmit={this.createNewComplianceRule}
+            />
+        })
     }    
 
 	render() {
@@ -85,6 +139,15 @@ FloatingActionButton = class FloatingActionButton extends React.Component {
 					onClick={this.createNewFacility} 
 					className="fab fab-2">
 						<i className="fa fa-building"></i>
+				</button>				
+				<button 
+					rel="tooltip"
+					data-toggle="tooltip" 
+					data-placement="left" 
+					title="Create new compliance rule"
+					onClick={()=>{this.handleCreateComplianceRuleClick()}} 
+					className="fab fab-3">
+						<i className="fa fa-check"></i>
 				</button>				
 			</div>
 		)